+++ packages/theme-preset/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @marigold/theme-preset
 
+## 1.3.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @marigold/system@7.5.3
+
+## 1.3.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @marigold/system@7.5.2
+
+## 1.3.6
+
+### Patch Changes
+
+- Updated dependencies [[`fba5e92`](https://github.com/marigold-ui/marigold/commit/fba5e92e8e70de0c3a65696cd2c462b6dfa7819e)]:
+  - @marigold/system@7.5.1
+
 ## 1.3.5
 
 ### Patch Changes